+++ CHANGELOG.md
@@ -6,6 +6,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.2.0 - UNRELEASED
+
+The plugin is now using the Freemarker template engine to generate source
+code. It defines two internal templates:
+
+* `keys` generates constants that holds the keys to access the properties file
+* `values` generate constants that holds the values from the properties file
+
+See documentation for details.
 
 ## 0.1.0 - 2021-10-17